Implement `Filterable`, add `mapWithAccum`+`when`
This commit adds two new functions to the library: `mapWithAccum`, a
function to help map over streams with respect to an accumulating value,
and `when`, which allows you to filter streams by other streams.

As well as this, the Filterable class has been added to replace a couple
of other utilities, such as `filter` and `mapMaybe`.