Add `withDelay` function
This commit adds the `withDelay` function to `FRP.Event.Time`. My
personal reason for using this was that, for every tick of the RAF
event, I wanted to progress a pre-scripted animation. So, if the
animation were to last for 1000ms, the `withDelay` stream let me check
the progress through the animation, and then eventually when it had
finished (i.e. to stop animating).