Add `mapWithAccum` to FRP.Event (#14)
* Add `mapWithAccum` to FRP.Event

The Event class now exports the `mapWithAccum` function. There was
originally concern that this would cause a huge composition chain, and
thus not be performant for events over a long period. However, I've
expanded the point-free folding function definition, so it's hopefully
now a bit clearer that it shouldn't happen. As always, though, @paf31,
do tell me if I'm being ignorant!

* Implement code review changes

s/mapWithAccum/mapAccum/g, and actually remembered to export it... >_<