Return the current time as a Number (#16)
The `withTime` event now returns `time` as a `Number`, rather than an
`Int`. This is probably sensible, as the current time since epoch in ms
is bigger than what we can store in a 32-bit int, so this gets truncated
as soon as it is used. This commit simply changes its representation to
a number, ensuring that it can remain as precise.