Add `gateBy` and `gate` to `FRP.Event.Time` (#21)
* Add `gateBy` and `gate` to `FRP.Event.Time`

Previously, the functions only existed for `Behavior`. This PR adds them
to `Event` as well.

* Fix comments from review

Put the functions in the wrong place. Oops. This commit moves the
`gate`-related functions from `FRP.Event.Time` to `FRP.Event.Class`, and
exports them.

* Final review changes, merging in blocking

As per the other PR, I've combined these two because there's some
interdependence. This implements the mentioned changes, and should now
be up-to-date with both PRs.

* Replace "initial gating value" with `Maybe`

Previously, the function required an initial value. This wasn't great if
you didn't have such a value to hand, and has hence been removed in
favour of a `Maybe` in the predicate. The result is much neater :)